requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-environ==0.12.0
celery==5.5.3
redis==7.0.0
channels-redis==4.3.0
drf-yasg==1.21.11
sentry-sdk==2.42.1
uvicorn[standard]==0.38.0
django-cors-headers==4.9.0
django-prometheus==2.4.1
django-celery-beat==2.8.1
firebase-admin==7.1.0
twilio==9.8.4
honcho==2.0.0
django-unfold==0.72.0